Move EqualityComparer.Create to TestCommon
